Replace psycopg2 with psycopg-2-binary in requirements
Psycopg2 package builds from source and requires some libraries,
which makes setting up development environment more difficult.

As per note in https://pypi.org/project/psycopg2-binary,
the binary package is a practical choice for development and
testing but in production it is advised to use the package built
from sources.

So this change should make thing easier locally, but in production
the psycopg2 package should be uncommented and used instead of
the binary one.