Change testing if social info is present in views.person
Since the relevant fields in profile have null=False by default
(null is database related) but are allowed to blank, empty values
are stored as empty strings, and thus can not be tested with
'x is None' to see if they have a value.

The django docs suggest to avoid using null on string-based fields
because having two possible ways of storing 'no data' is ambiguous.