Adjust requirements file to behave better in production
We want psycopg in development and psycopg-binary in testing
because the former is harder to build, see notes in
https://pypi.org/project/psycopg2/

Previously, the psycopg was commented out and psycopg-binary was
being used. This favoured testing, but made deployment harder
since we would have to edit the file every time we wanted to
run updates with pip install -r.

As a partial solution, I set psycopg to be used on Linux and
psycopg-binary on MacOS. Now, production will use the right one,
development on Mac will use the right one, and manual edits will
only be needed for development on linux.