Fix placement of html tags
OCD kicked in. This commit introduces following changes:
- Put opening <p> tags after <h3>...</h3> in index & news
- Put <b> tags inside <a> tags in team (bolded name & position)
- Remove commas separating attributes in person (github logo etc.)

Only the first change is functional: now each news item will be placed
inside a paragraph and will get formatted according to css specification
for a paragraph (most notably, will have line spacing). Previously, the
header cancelled the paragraph, and the closing p tag was left hanging.
This way seems to be closer to original intentions, and if the effect is
not desired, changes should be made to css.

Other two changes are for pure syntactic correctness, though page displayed
correctly without them.

Now that the <p> tags are properly placed, we will have extraneous opening
tags if somebody places them in the news item. Shouldn't break the way things
are displayed but we may wish to create a filter for that.