Add dropptag filter and use it in index, news & person
The new filter drops opening and closing <p> tags from a piece of
text. I think it's good to have it because we have the <p> tags in
html templates, and we sometimes put them and sometimes don't when
we write content through admin panel.

Of course two opening or closing tags (<p><p>) won't crash anything,
but they look annoying in source ;).